+++ src/libs/tigon/Representation/Functions/CTP/CTP1.cpp
@@ -66,7 +66,7 @@ void CTP1::evaluate(const TVector<IElementSPtr> &inputs, const TVector<IElementS
         TVector<double> x = IElementVecToRealVec(inputs);
         TVector<double> y = CTP::CTP1(x);
 
-        for(int i=0; i<outputs.size(); i++) {
+        for(size_t i=0; i<outputs.size(); i++) {
             outputs[i]->defineValue(y[i]);
         }
     } else {
