+++ src/libs/tigon/Representation/Functions/CTP/CTP2.cpp
@@ -71,13 +71,14 @@ CTP2::~CTP2()
 
 }
 
-void CTP2::evaluate(const TVector<IElementSPtr> &inputs, const TVector<IElementSPtr>& outputs)
+void CTP2::evaluate(const TVector<IElementSPtr>& inputs,
+                    const TVector<IElementSPtr>& outputs)
 {
     if((inputs.size() == TP_nInputs()) && (outputs.size() == TP_nOutputs())) {
         TVector<double> x = IElementVecToRealVec(inputs);
         TVector<double> y = CTP::CTP2(x);
 
-        for(int i=0; i<outputs.size(); i++) {
+        for(size_t i=0; i<outputs.size(); i++) {
             outputs[i]->defineValue(y[i]);
         }
     } else {
