+++ src/libs/tigon/Representation/Functions/CTP/CTP4.cpp
@@ -72,13 +72,14 @@ CTP4::~CTP4()
 
 }
 
-void CTP4::evaluate(const TVector<IElementSPtr> &inputs, const TVector<IElementSPtr>& outputs)
+void CTP4::evaluate(const TVector<IElementSPtr> &inputs,
+                    const TVector<IElementSPtr> &outputs)
 {
     if((inputs.size() == TP_nInputs()) && (outputs.size() == TP_nOutputs())) {
         TVector<double> x = IElementVecToRealVec(inputs);
         TVector<double> y = CTP::CTP4(x);
 
-        for(int i=0; i<outputs.size(); i++) {
+        for(size_t i=0; i<outputs.size(); i++) {
             outputs[i]->defineValue(y[i]);
         }
     } else {
