+++ src/libs/tigon/Representation/Functions/CTP/CTP5.cpp
@@ -70,13 +70,14 @@ CTP5::~CTP5()
 
 }
 
-void CTP5::evaluate(const TVector<IElementSPtr> &inputs, const TVector<IElementSPtr>& outputs)
+void CTP5::evaluate(const TVector<IElementSPtr> &inputs,
+                    const TVector<IElementSPtr> &outputs)
 {
     if((inputs.size() == TP_nInputs()) && (outputs.size() == TP_nOutputs())) {
         TVector<double> x = IElementVecToRealVec(inputs);
         TVector<double> y = CTP::CTP5(x);
 
-        for(int i=0; i<outputs.size(); i++) {
+        for(size_t i=0; i<outputs.size(); i++) {
             outputs[i]->defineValue(y[i]);
         }
     } else {
